requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.36
alembic==1.13.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0
PyMySQL==1.1.0
email-validator==2.1.1
python-dotenv==1.0.1
pytest==8.0.2
httpx==0.26.0
